Support wider thread ID for Macos/x64. Also remove the sigsegv handler, which is not helping when debugging.
git-svn-id: https://llvm.org/svn/llvm-project/vmkit/trunk@139081 91177308-0d34-0410-b5e6-96231b3b80d8

+++ lib/Mvm/CommonThread/ObjectLocks.cpp
@@ -240,7 +240,7 @@ FatLock* ThinLock::getFatLock(gc* object, LockSystem& table) {
   }
 }
 
-void FatLock::acquireAll(gc* object, uint32 nb) {
+void FatLock::acquireAll(gc* object, uintptr_t nb) {
   assert(associatedObject == object);
   llvm_gcroot(object, 0);
   internalLock.lockAll(nb);
